feat: reset variable values on network destroy (may return to pool)

+++ Scripts/Gameplay/CharacterEntity/MonsterCharacterEntity/BaseMonsterCharacterEntity_Cleanup.cs
@@ -1,23 +1,32 @@
+using UnityEngine;
+
 namespace MultiplayerARPG
 {
     public partial class BaseMonsterCharacterEntity
     {
-        public override void Clean()
+        public override void Clean(bool isObjectDestroyed)
         {
-            base.Clean();
-            characterDatabase = null;
-            faction = null;
-            summoner = null;
+            base.Clean(isObjectDestroyed);
+            if (isObjectDestroyed)
+            {
+                characterDatabase = null;
+                faction = null;
+                InstantiatedObjects.DestroyAndNullify();
+                InstantiatedObjects.Clear();
+                _isObjectsInstantiated = false;
+            }
             SpawnArea = null;
             SpawnPrefab = null;
 #if !DISABLE_ADDRESSABLES
             SpawnAddressablePrefab = null;
 #endif
+            SpawnLevel = 0;
+            SpawnPosition = Vector3.zero;
+            _isDestroyed = false;
             _looters?.Clear();
             _droppingItems?.Clear();
-
-            InstantiatedObjects.DestroyAndNullify();
-            InstantiatedObjects.Clear();
+            _killedReward = null;
+            _lastTeleportToSummonerTime = 0f;
         }
     }
 }
