+++ Scripts/Gameplay/Npcs/NpcEntity_Cleanup.cs
@@ -2,23 +2,26 @@ namespace MultiplayerARPG
 {
     public partial class NpcEntity
     {
-        public override void Clean()
+        public override void Clean(bool isObjectDestroyed)
         {
-            base.Clean();
-            startDialog = null;
-            graph = null;
-            characterUiTransform = null;
-            miniMapUiTransform = null;
-            questIndicatorContainer = null;
-            if (_uiNpcEntity != null)
-                Destroy(_uiNpcEntity.gameObject);
-            _uiNpcEntity = null;
-            if (_questIndicator != null)
-                Destroy(_questIndicator.gameObject);
-            _questIndicator = null;
-
-            InstantiatedObjects.DestroyAndNullify();
-            InstantiatedObjects.Clear();
+            base.Clean(isObjectDestroyed);
+            if (isObjectDestroyed)
+            {
+                startDialog = null;
+                graph = null;
+                characterUiTransform = null;
+                miniMapUiTransform = null;
+                questIndicatorContainer = null;
+                if (_uiNpcEntity != null)
+                    Destroy(_uiNpcEntity.gameObject);
+                _uiNpcEntity = null;
+                if (_questIndicator != null)
+                    Destroy(_questIndicator.gameObject);
+                _questIndicator = null;
+                InstantiatedObjects.DestroyAndNullify();
+                InstantiatedObjects.Clear();
+                _isObjectsInstantiated = false;
+            }
         }
     }
 }
